Moving NITFDetails definition to the appropriate place
